chore: :zap: update ping request to include timezone in response
The code changes modify the `app.controller.ts` and `app.service.ts` files in the `be` directory. The changes include updating the `getPing` method in the `AppController` class to accept a `Request` object in addition to the `Response` object. The `getPing` method in the `AppService` class is also updated to accept a `Request` object. The `toZonedTime` function is called with the `req.body['tz']` value or the `process.env['TZ']` value as the timezone parameter.

Based on the recent repository commits, it seems that there have been updates related to the ping functionality and user management.